Added script to submit jobs to AWE.  The script requires a metadata file and sequence file already loaded into Shock and a configuration file (also added to the repo here) and a pipeline/workflow file (also added to the repo here).
Additionally, the user must be authenticated by logging into IRIS (KB_AUTH_USER_ID should be set in ENV) or by
setting a globus online user and password in the awe.ini config file.

Finally, the metadata file has to pass validation and either the jobname must match a metagenome_name in the
metadata file, or the file name must match a file_name in the metadata file for successful AWE submission.